requirements.txt
Flask==3.0.0
confluent-kafka==2.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
werkzeug==3.0.1
